HOTFIX: fixed infiniate infinateloop
Fixed infinate loop where when we needed to realloc whilst reading the
rest of a line we found in "restOfLineStartingWith". we where never
updating the buffer length so we would realloc but not store the new
size and thus when it reentered the for loop it would still think it was
still at the end of the buffer and try to realloc again.